requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.4  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==1.2.0  # https://github.com/mher/flower
django-annoying==0.10.6
python-keycloak==3.9.1
loguru==0.7.0
djangorestframework-datatables==0.7.0
requests==2.30.0
tqdm==4.65.0
pydantic==2.4.2
kubernetes==26.1.0
markdown==3.4.3
docker==6.1.2
django-filter==23.2
owlready2==0.44
pandas==2.2.2
lxml==4.9.3

# Django
# ------------------------------------------------------------------------------
django==4.1.9  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.57.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==3.14.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.2  # https://github.com/tfranzel/drf-spectacular
django-webpack-loader==1.8.1  # https://github.com/django-webpack/django-webpack-loader

httpx==0.25.0
django_components==0.29
drf-extensions==0.7.1
django-active-link==0.1.8
python-magic==0.4.27
django-constance==3.1.0
websocket-client==1.8.0
rel==0.4.9.19
web3==7.2.0
tenacity==9.0.0
